requirements.txt
pydantic==1.8.2
textblob==0.15.3
nltk==3.6.1
PyMySQL==1.0.2
fastapi==0.68.1
python-dotenv==0.19.0
scikit_learn==1.0
